kvm: Handle a case when interrupts are enabled in the kernel space
This can happen when goruntime switches on a goroutine which
has been suspended in the host mode. When goruntime restores
a goroutine state, it calls popf which restores rlags from
a stack. If the routine has been suspended in the host mode,
its rflag will have the IF bit and restoring these flags will
enables interrupts.

This patch instroduces a hacky way to handle this case. gVisor doesn't
allow to change a value of the GS register from user processes. This
means that we can set it for userspace processes and use it to determin
in what context a system call instruction has been called.