kvm: limit a number of vcpu-s
GOMAXPROCS() is a number of "CPUs" that is avaliable for us.
Out goal is avoid battles for vCPUs.

Goi-runtime can execute up to GOMAXPROCS threads concurrently.

And let's think that the user applications  wants to reach a maximum
performace and so limie a number of concurrently running threads by the
number of avaliable CPUs.