[POC] kvm: map the entire sentry address space to the vm
It is required to avoid trapping mmap syscalls.

Right now, we limit the sentry address space to the size of a vm
physical address space. With this change we need to limit it even more.
The problem here is that the kernel allocates one by per page, so event
for the 39 bits address space, the overhead is 128MB. In the kernel,
this memory is allocated by vmalloc, so it is time consuming too.